# perf/limb-decompose-bytes